(PC-17092)[API] chore: move educational commands in a commands file
